+++ src/components/WinrateGraph.js
@@ -12,7 +12,14 @@ class WinrateGraph extends Component {
 
         this.state = {
             height: setting.get('view.winrategraph_height'),
+            invert: setting.get('view.winrategraph_invert')
         }
+
+        setting.events.on('change', ({key, value}) => {
+            if (key === 'view.winrategraph_invert') {
+                this.setState({invert: value})
+            }
+        })
 
         this.handleMouseDown = evt => {
             this.mouseDown = true
@@ -25,11 +32,12 @@ class WinrateGraph extends Component {
         }
     }
 
-    shouldComponentUpdate({width, currentIndex, data}, {height}) {
+    shouldComponentUpdate({width, currentIndex, data}, {height, invert}) {
         return width !== this.props.width
             || currentIndex !== this.props.currentIndex
             || data[currentIndex] !== this.props.data[currentIndex]
             || height !== this.state.height
+            || invert !== this.state.invert
     }
 
     componentDidMount() {
@@ -65,11 +73,11 @@ class WinrateGraph extends Component {
 
     }
 
-    render({width, currentIndex, data}) {
+    render() {
+        let {width, currentIndex, data} = this.props
+        let {invert} = this.state
 
-        let inv = setting.get('view.winrategraph_invert')
-
-        if (inv) {
+        if (invert) {
             data = data.map(x => x == null ? null : 100 - x)
         }
 
@@ -102,7 +110,7 @@ class WinrateGraph extends Component {
                 // Draw background
 
                 h('defs', {},
-                    h('linearGradient', {id: 'bgGradient', x1: 0, y1: inv ? 1 : 0, x2: 0, y2: inv ? 0 : 1},
+                    h('linearGradient', {id: 'bgGradient', x1: 0, y1: invert ? 1 : 0, x2: 0, y2: invert ? 0 : 1},
                         h('stop', {
                             offset: '0%',
                             'stop-color': 'white',
@@ -127,9 +135,9 @@ class WinrateGraph extends Component {
 
                                 if (instructions.length === 0) return ''
 
-                                return `M ${instructions[0][0]},${inv ? 0 : 100} `
+                                return `M ${instructions[0][0]},${invert ? 0 : 100} `
                                     + instructions.map(x => `L ${x.join(',')}`).join(' ')
-                                    + ` L ${instructions.slice(-1)[0][0]},${inv ? 0 : 100} Z`
+                                    + ` L ${instructions.slice(-1)[0][0]},${invert ? 0 : 100} Z`
                             })()
                         })
                     )