+++ src/components/drawers/PreferencesDrawer.js
@@ -27,9 +27,9 @@ class PreferencesItem extends Component {
             onChange(Object.assign({checked}, this.props))
         }
 
-        setting.events.on('change', ({key}) => {
+        setting.events.on('change', ({key, value}) => {
             if (key === this.props.id) {
-                this.setState({checked: setting.get(key)})
+                this.setState({checked: value})
             }
         })
     }
@@ -203,9 +203,9 @@ class PathInputItem extends Component {
             })
         }
 
-        setting.events.on('change', ({key}) => {
+        setting.events.on('change', ({key, value}) => {
             if (key === this.props.id) {
-                this.setState({value: setting.get(key)})
+                this.setState({value: value})
             }
         })
     }
@@ -319,9 +319,9 @@ class ThemesTab extends Component {
             })
         }
 
-        setting.events.on('change', ({key}) => {
+        setting.events.on('change', ({key, value}) => {
             if (key === 'theme.current') {
-                this.setState({currentTheme: setting.get(key)})
+                this.setState({currentTheme: value})
             }
         })
     }